Add initial basic binding
This adds a simple binding using BuckleScript to the fetch function. It
also adds a function to be consumed by library users on top of this that
converts the promise from the JS `fetch` function to a Reason Promise
that with a result object.

Some raw JavaScript is included to help type the error that the promise
returns. This could be implemented by copying the implementation of
`toResult` and doing the error conversion in the same step. However, the
performance gains are probably neglible (compared to the network I/O
performed) and it's less readable.

This also includes the initial set-up for the init function. There and
in the response object is where most work will still need to be done for
user ergonomics.

It's up for debate whether to rename `init` to `options` because that
may be clearer to people coming to fetch for the first time. I believe
`init` was chosen by the spec because they consistently talk about
"initialising" the fetch/request/headers.