Drop Node.js as goal for binding support
`fetch` itself is not directly supported on Node.js. This would mean
that to support `fetch` on Node.js we'd have to pick an interop library.
Most options that can easily be found right now seem quite outdated.

To make it easier on ourselves, we'll focus on the browser for now and
see how difficult it is to add something for Node later.

People involved in the Node.js project seem to be against directly
integrating `fetch` in Node.js because the constraints of a browser vs a
server are inherently different (e.g. CORS necessity). With this in mind
it may be good to pass the responsibility of decisions about these
things on to the implementing developer.

A discussion about this can be found here: https://github.com/nodejs/node/issues/19393