Fix bug allowing AbortController to be constructed from signal
An AbortController has a signal but it's more than a struct with a
signal. To prevent people from being able to pass off `{ signal }` as an
AbortController and breaking their program, the type is marked as
private. This still allows other modules to use it as input/output types
for functions or other types but doesn't allow other modules to
construct the type directly.