Implement partial Response type
This adds a module that covers the `Response` class of JavaScript. Some
of the utility methods are implemented already. For now we stick to
`json` and `text` as these are the most useful methods for the author's
current use-case and they also map most easily to existing types.

Some extra error handling types are added to deal with the possibility
of trying to read the body of a response multiple times.