Begin implementing fetch Headers
This is the start of implementing header support for fetch requests and
responses. A decision still needs to be made on how rigid to implement
these.

Consuming the Twitch API already shows that real-world usage could
require using non-standard headers that are not prefixed with `X-`. This
would make it infeasible to implement strict header validation
(requiring any non `X-` prefixed header to be validated).

A compromise could be made adding 'unsafe' methods in the future. For
now there is no header validation done at all and header access can
cause exceptions, so this will change in the future.